Refactor completers to separate file
